backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
python-multipart==0.0.6
requests==2.31.0
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
pydantic==2.5.2
pydantic-settings==2.1.0
openai==1.6.1
google-auth==2.25.2
httpx==0.25.2
pymupdf==1.23.6
anthropic==0.7.0
redis==5.0.1
pycryptodome==3.19.0
tenacity==8.2.3
